+++ mne/viz/tests/test_topomap.py
@@ -60,6 +60,12 @@ def test_plot_topomap():
     res = 16
     evoked = read_evokeds(evoked_fname, 'Left Auditory',
                           baseline=(None, 0))
+    # Test animation
+    _, anim = evoked.animate_topomap(ch_type='grad', times=[0, 0.1],
+                                     butterfly=False)
+    anim._func(1)  # _animate has to be tested separately on 'Agg' backend.
+    plt.close('all')
+
     ev_bad = evoked.pick_types(meg=False, eeg=True, copy=True)
     ev_bad.pick_channels(ev_bad.ch_names[:2])
     ev_bad.plot_topomap(times=ev_bad.times[:2] - 1e-6)  # auto, should plot EEG